cast for enums

+++ programs/pluto/pam_conv.c
@@ -332,7 +332,7 @@ void *pam_thread(void *parg)
   pthread_mutex_destroy(&thread_run_m);
 
   libreswan_log("XAUTH: #%lu: PAM thread completed pam_do_state == %s pam_state == %s", _serialno,
-      pam_result_state_enum[_pam_do_state], pam_state_enum[_pam_state] );
+      pam_result_state_enum[(int)_pam_do_state], pam_state_enum[(int)_pam_state] );
   usleep(200000);
   return NULL;
 }
